amending changes from main into experimental

+++ composeApp/src/commonMain/kotlin/utils/StepCounter.kt
@@ -4,7 +4,7 @@ package utils
 expect abstract class PlatformContext
 
 expect class StepCounter(context: PlatformContext) {
-    fun startListening(stepsGoal: Int, onGoalAchieved: () -> Unit)
+    fun startListening(stepsGoal: Int, onGoalAchieved: (Int) -> Unit)
     fun stopListening()
     var stepCount: Int
 }
